List Amiga Basic reserved words.

Obtain a list of reserved Amiga Basic words. These words are not allowed as names of variables or labels in Amiga Basic.

Usage

AmigaBasic.reserved()

Value

Returns a vecor of character strings of reserved Amiga Basic words.

Details

This function will return a full list of reserved Amiga Basic words. This list does not serve as a manual for basic (for that purpose consult external resources). This list is meant to consult when choosing label names in Amiga Basic code. These reserved words are not allowed as names.

See also

Author

Pepijn de Vries

Examples

AmigaBasic.reserved()
#>   [1] "'"               "*"               "+"               "-"              
#>   [5] "/"               "<"               "="               ">"              
#>   [9] "ABS"             "ALL"             "AND"             "APPEND"         
#>  [13] "AREA"            "AREAFILL"        "AS"              "ASC"            
#>  [17] "ATN"             "BASE"            "BEEP"            "BREAK"          
#>  [21] "CALL"            "CDBL"            "CHAIN"           "CHDIR"          
#>  [25] "CHR$"            "CINT"            "CIRCLE"          "CLEAR"          
#>  [29] "CLNG"            "CLOSE"           "CLS"             "COLLISION"      
#>  [33] "COLOR"           "COMMON"          "CONT"            "COS"            
#>  [37] "CSNG"            "CSRLIN"          "CVD"             "CVI"            
#>  [41] "CVL"             "CVS"             "DATA"            "DATE$"          
#>  [45] "DECLARE"         "DEF"             "DEFDBL"          "DEFINT"         
#>  [49] "DEFLNG"          "DEFSNG"          "DEFSTR"          "DELETE"         
#>  [53] "DIM"             "ELSE"            "ELSEIF"          "END"            
#>  [57] "EOF"             "EQV"             "ERASE"           "ERL"            
#>  [61] "ERR"             "ERROR"           "EXIT"            "EXP"            
#>  [65] "FIELD"           "FILES"           "FIX"             "FN"             
#>  [69] "FOR"             "FRE"             "FUNCTION"        "GET"            
#>  [73] "GOSUB"           "GOTO"            "HEX$"            "IF"             
#>  [77] "IMP"             "INKEY$"          "INPUT"           "INSTR"          
#>  [81] "INT"             "KILL"            "LBOUND"          "LEFT$"          
#>  [85] "LEN"             "LET"             "LIBRARY"         "LINE"           
#>  [89] "LLIST"           "LOAD"            "LOC"             "LOCATE"         
#>  [93] "LOF"             "LOG"             "LPOS"            "LPRINT"         
#>  [97] "LSET"            "MENU"            "MERGE"           "MID$"           
#> [101] "MKD$"            "MKI$"            "MKL$"            "MKS$"           
#> [105] "MOD"             "MOUSE"           "NAME"            "NEW"            
#> [109] "NEXT"            "NOT"             "OBJECT.AX"       "OBJECT.AY"      
#> [113] "OBJECT.CLIP"     "OBJECT.CLOSE"    "OBJECT.HIT"      "OBJECT.OFF"     
#> [117] "OBJECT.ON"       "OBJECT.PLANES"   "OBJECT.PRIORITY" "OBJECT.SHAPE"   
#> [121] "OBJECT.START"    "OBJECT.STOP"     "OBJECT.VX"       "OBJECT.VY"      
#> [125] "OBJECT.X"        "OBJECT.Y"        "OCT$"            "OFF"            
#> [129] "ON"              "OPEN"            "OPTION"          "OR"             
#> [133] "OUTPUT"          "PAINT"           "PALETTE"         "PATTERN"        
#> [137] "PEEK"            "PEEKL"           "PEEKW"           "POINT"          
#> [141] "POKE"            "POKEL"           "POKEW"           "POS"            
#> [145] "PRESET"          "PRINT"           "PSET"            "PTAB"           
#> [149] "PUT"             "RANDOMIZE"       "READ"            "REM"            
#> [153] "RESET"           "RESTORE"         "RESUME"          "RETURN"         
#> [157] "RIGHT$"          "RND"             "RSET"            "RUN"            
#> [161] "SADD"            "SAVE"            "SAY"             "SCREEN"         
#> [165] "SCROLL"          "SGN"             "SHARED"          "SIN"            
#> [169] "SLEEP"           "SOUND"           "SPACE$"          "SPC"            
#> [173] "SQR"             "STATIC"          "STEP"            "STICK"          
#> [177] "STOP"            "STR$"            "STRIG"           "STRING$"        
#> [181] "SUB"             "SWAP"            "SYSTEM"          "TAB"            
#> [185] "TAN"             "THEN"            "TIME$"           "TIMER"          
#> [189] "TO"              "TRANSLATE$"      "TROFF"           "TRON"           
#> [193] "UBOUND"          "UCASE$"          "USING"           "VAL"            
#> [197] "VARPTR"          "WAIT"            "WAVE"            "WEND"           
#> [201] "WHILE"           "WIDTH"           "WINDOW"          "WRITE"          
#> [205] "XOR"             "\\"              "^"