A list of special display modes

A list of special display modes on the Amiga and corresponding raw keys.

Format

A data.frame with 2 columns:

The column named 'mode': a factor reflecting a display mode, monitor or bitwise mask

The column named 'code': vector of 4 raw values as used by the Amiga to reflect specific display modes

Details

This table show specific special display modes and to which Amiga monitors they relate. The raw codes can be used to interpret specific display modes as listed in amiga_display_modes(). This information is used to interpret IFFChunk() objects of type 'CAMG'. It is also used to interpret ILBM images and creating IFF files from raster images.

References

https://wiki.amigaos.net/wiki/Display_Database#ModeID_Identifiers

http://amigadev.elowar.com/read/ADCD_2.1/AmigaMail_Vol2_guide/node00FD.html

Examples

data("amiga_display_keys")