Convert AmigaFFH objects into raw data

Convert AmigaFFH objects into raw data, as they would be stored in the Commodore Amiga's memory or files.

Usage

# S3 method for class 'AmigaBasic'
as.raw(x, ...)

# S3 method for class 'AmigaBasicShape'
as.raw(x, ...)

# S3 method for class 'AmigaBasicBMAP'
as.raw(x)

# S3 method for class 'AmigaBitmapFont'
as.raw(x, ...)

# S3 method for class 'AmigaBitmapFontSet'
as.raw(x, ...)

# S3 method for class 'AmigaTimeVal'
as.raw(x, ...)

# S4 method for class 'hardwareSprite'
as.raw(x)

# S4 method for class 'IFFChunk'
as.raw(x)

# S3 method for class 'IFF.ANY'
as.raw(x, ...)

# S3 method for class 'SysConfig'
as.raw(x, ...)

# S3 method for class 'AmigaIcon'
as.raw(x, ...)

Arguments

x

An AmigaFFH object that needs to be converted into raw data. See usage section for all supported objects.

...

Arguments passed on to IFFChunk-method() when x is of class IFF.ANY.

Value

Returns a vector of raw data based on x.

Details

Objects originating from this package can in some cases be converted into raw data, as they would be stored on an original Amiga. See the usage section for the currently supported objects.

Not all information from x may be included in the raw data that is returned, so handle with care.

As this package grows additional objects can be converted with this method.

See also

Author

Pepijn de Vries

Examples

## read an IFF file as an IFFChunk object:
example.iff <- read.iff(system.file("ilbm8lores.iff", package = "AmigaFFH"))

## This will recreate the exact raw data as it was read from the file:
example.raw <- as.raw(example.iff)