Get available font sizes from an AmigaBitmapFontSet

Get available font sizes (height) from an AmigaBitmapFontSet() in pixels.

Usage

availableFontSizes(x)

Arguments

x

An AmigaBitmapFontSet() for which the available font sizes (height) in number of pixels need to be obtained.

Value

Returns a vector of numeric values specifying the available font sizes (height in pixels) for x.

Details

An AmigaBitmapFontSet() can hold bitmaps of multiple font sizes. Use this function to obtain the available size from such a set.

See also

Author

Pepijn de Vries

Examples

data(font_example)

## The example font holds two font sizes (8 and 9):
availableFontSizes(font_example)
#> [1] 8 9