Check Amiga Basic label/variable names for validity

Check Amiga Basic label/variable names for validity

Usage

check.names.AmigaBasic(x, ...)

Arguments

x

A vector of character strings that need to be checked

...

Currently ignored.

Value

A data.frame with logical values with the same number of rows as the length of x. Columns in the data.frame corresponds with the criteria listed in the details. FALSE for invalid names.

Details

Names for variables and labels should adhere to the following rules in Amiga Basic:

Length of the names should be in the range of 1 up to 255 character

Names cannot be AmigaBasic.reserved() words

Names should only contain alphanumeric characters or periods and should not contain special characters (i.e., reserved for type definition, such as dollar- or percentage sign)

Names should not start with a numeric character

This function tests names against each of these criteria.

See also

Author

Pepijn de Vries

Examples

## These are valid names in Amiga Basic:
check.names.AmigaBasic(c("Foo", "Bar"))
#>     duplicated length reserved characters start
#> Foo      FALSE  FALSE    FALSE      FALSE FALSE
#> Bar      FALSE  FALSE    FALSE      FALSE FALSE

## Reserved words and repeated names are not allowed:

check.names.AmigaBasic(c("Print", "Foo", "Foo"))
#>       duplicated length reserved characters start
#> Print      FALSE  FALSE     TRUE      FALSE FALSE
#> Foo        FALSE  FALSE    FALSE      FALSE FALSE
#> Foo.1       TRUE  FALSE    FALSE      FALSE FALSE