An example object for the AmigaBitmapFontSet class

An example object for the AmigaBitmapFontSet() class used in examples throughout this package. It also contains a nested AmigaBitmapFont() class objects, which can be obtain by using getAmigaBitmapFont(font_example, 9).

Format

font_example is an AmigaBitmapFontSet() object. For details see the object class documentation.

Details

The font_example contains a font that was designed as an example for this package. It holds bitmap glyphs for 8 and 9 pixels tall characters.

See also

Examples

data("font_example")