Extract or replace variable and label names from Amiga Basic scripts

In the binary Amiga Basic files, names for labels and variables in the code are stored at the end of the file. In the encoded there is only a pointer to the index of the name in that list. Use this function to list, select or replace names included in the code

Usage

# S3 method for class 'AmigaBasic'
names(x)

# S3 method for class 'AmigaBasic'
names(x) <- value

Arguments

x

An AmigaBasic()-class object for which to obtain or change variable and/or label names

value

A (vector of) character string of desired replacement variable/label names.

Value

A vector of character strings with label and variable names in the basic script. In case of the replacement method a AmigaBasic()-class with replaced names is returned.

Details

Make sure that variable and label names are valid for the basic script (see check.names.AmigaBasic).

See also

Author

Pepijn de Vries

Examples

## Let's create some Basic code with labels and variables:
bas <- as.AmigaBasic(c(
  "REM - This will loop forever...",
  "my.label:",
  "  my.variable% = 0",
  "  WHILE my.variable% < 10",
  "    my.variable% = my.variable% + 1",
  "  WEND",
  "  GOTO my.label"
))

## list the names in the script above:
names(bas)
#> [1] "my.label"    "my.variable"

## change the first name:
names(bas)[1] <- "better.label"