Playing Amiga audio data

A wrapper for tuneR package's tuneR::play() routine. Allowing it to play Amiga audio (for instance stored in an 8SVX Interchange File Format).

Usage

# S4 method for class 'ANY'
play(object, player = NULL, ...)

# S4 method for class 'IFFChunk'
play(object, player = NULL, ...)

Arguments

object

An IFFChunk-class() object that needs to be played. The IFFChunk() should be of type FORM, containing an 8SVX chunk, or an 8SVX itself. object can also be of class IFF.FORM or IFF.8SVX. See tuneR::play() for other objects that can be played.

player

Path to the external audio player. See tuneR::play() for more details.

...

Arguments passed onto the tuneR play() routine.

Value

Returns a list of data returned by tuneR's tuneR::play(), for which the output is undocumented.

Details

A wrapper for tuneR package's tuneR::play() routine. It will try to play audio using an external audio player. When 8SVX audio is played, each octave is played separately. When a FORM container contains multiple 8SVX samples, they are also played successively.

Note that a separate package is developed to interpret and play ProTracker modules and samples (ProTrackR()).

Author

Pepijn de Vries

Examples

## First get an audio sample from the ProTrackR package
snare.samp <- ProTrackR::PTSample(ProTrackR::mod.intro, 2)

## Coerce it into an IFFChunk object:
snare.iff <- WaveToIFF(snare.samp)

## Play the 8SVX sample:
if (interactive()) {
  play(snare.iff)
}