Coerce raw data into an AmigaBitmapFontSet class object

AmigaBitmapFontSet() objects are comprehensive representations of binary Amiga font files (*.font). Use this function to convert raw data from such a file to an AmigaBitmapFontSet object.

Usage

rawToAmigaBitmapFontSet(x, file)

Arguments

x

A vector of raw data that needs to be converted into an AmigaBitmapFontSet().

file

The raw version of the AmigaBitmapFontSet() does not contain the nested font bitmap images. In order to correctly construct an AmigaBitmapFontSet() the file location of the original *.font file is required in order to read and include the font bitmap image information. file should thus be a character string specifying the file location of the *.font file.

Value

Returns an AmigaBitmapFontSet() object.

Details

This function converts raw data as stored in *.font files. The function also needs the file location, in order to load the nested bitmap images for each font height. This function is effectively the inverse of as.raw().

See also

Author

Pepijn de Vries

Examples

data(font_example)

## First create raw font set data. Note that this raw data
## does not include the nested font bitmap images.
fontset.raw <- as.raw(font_example)

## Therefore it is necesary to have the entire font stored as files:
write.AmigaBitmapFontSet(font_example, tempdir())

font.restored <- rawToAmigaBitmapFontSet(fontset.raw, file.path(tempdir(), "AmigaFFH.font"))