Coerce raw data into an AmigaIcon class object

AmigaIcon() objects are comprehensive representations of binary Amiga Workbench icon files (*.info). Use this function to convert raw data from such a file to an AmigaIcon() object.

Usage

rawToAmigaIcon(x, palette = NULL)

Arguments

x

A vector of raw data that needs to be converted into an S3 AmigaIcon() class object.

palette

Provide a palette (vector of colours) for the icon bitmap image. When set to NULL (default) the standard Amiga Workbench palette will be used.

Value

Returns an AmigaIcon() class object based on x.

Details

Icons files (*.info) were used as a graphical representations of files and directories on the Commodore Amiga. This function will convert the raw data from such files into a more comprehensive names list (see AmigaIcon()). Use as.raw() to achieve the inverse.

See also

Author

Pepijn de Vries

Examples

## generate a simple AmigaIcon object:
icon <- simpleAmigaIcon()

## convert it into raw data:
icon.raw <- as.raw(icon)

## convert the raw data back into an icon:
icon.restored <- rawToAmigaIcon(icon.raw)