Read an AmigaBitmapFont class object from a file

Amiga Font Bitmaps of distinctive font heights are stored in separate files, which in combination form a font collection or set. This function can be used to read a specific bitmap from a set and returns it as an AmigaBitmapFont() class object.

Usage

read.AmigaBitmapFont(file, ...)

Arguments

file

The file name of a font subset is usually simply a numeric number indicating the font height in pixels. Use file as a character string representing that file location.

...

Arguments passed on to rawToAmigaBitmapFont().

Value

Returns an AmigaBitmapFont() object read from the specified file.

Details

Individual font bitmaps are stored in a font's subdirectory where the file name is usually equal to the font height in pixels. This function will read such a font bitmap file and return it as an AmigaBitmapFont() class object. It can also read such files from virtual disks (adf_file_con()) objects, but that requires the adfExplorer package to be installed.

See also

Author

Pepijn de Vries

Examples

data(font_example)

## Let's store the example font first:
write.AmigaBitmapFontSet(font_example, tempdir())

## Now read a specific subset from the font files:
font.sub <- read.AmigaBitmapFont(file.path(tempdir(), "AmigaFFH", "9"))

## The same can be done with a virtual Amiga disk. The following
## examples require the 'adfExplorer' package.
if (requireNamespace("adfExplorer")) {
  library("adfExplorer")
  virtual_disk_file <- tempfile(fileext = ".adf") |>
    create_adf_device(write_protected = FALSE) |>
    prepare_adf_device("font_disk") |>
    make_adf_dir("FONTS")
  
  dest <- virtual_path(virtual_disk_file, "DF0:FONTS")
  write.AmigaBitmapFontSet(font_example, dest)
  font.read <- read.AmigaBitmapFont(
    virtual_path(virtual_disk_file, "DF0:FONTS/AmigaFFH/9")
  )
  close(virtual_disk_file)
}