Write an AmigaBitmapFont(set) file

Functions to write AmigaBitmapFont() and AmigaBitmapFontSet() class objects to files.

Usage

write.AmigaBitmapFont(x, file)

write.AmigaBitmapFontSet(x, path = getwd())

Arguments

x

Respectively an AmigaBitmapFont() or a AmigaBitmapFontSet() object depending on which of the write-functions is called. This is the object that will be written to the specified file.

file

A character string specifying the file location to which x (an AmigaBitmapFont() object) needs to be written. It is common practice on the Amiga to use the font height in pixels as file name.

path

A character string specifying the path where x (an AmigaBitmapFontSet() object) needs to be stored. The filename for the font set will be extracted from x using fontName() followed by the *.font extension. A subdirectory will be created with the same name (without the extension) if it doesn't already exists. In this subdirectory all the nested AmigaBitmapFont() objects are stored.

Value

Invisibly returns the result of the call of close to the file connection.

Details

AmigaBitmapFontSet() class objects are written to a *.font file. The filename used for this purpose is obtained from the object itself using fontName(). In addition, a subdirectory is created automatically (when it doesn't already exist) to which al the separate bitmap images for each font height are written to individual files.

AmigaBitmapFont() class objects can also be written to a file. In order to use it on a Commodore Amiga or emulator, it is better to embed the font bitmap in a font set (using c()) and write the set to corresponding files.

See also

Author

Pepijn de Vries

Examples

## obtain a bitmap font set:
data(font_example)

## write the font set to their files. The file name
## is extracted from the font object, so you only have
## to provide the path:
write.AmigaBitmapFontSet(font_example, tempdir())

## extract a font bitmap:
font <- getAmigaBitmapFont(font_example, 9)

## and write it to the temp dir:
write.AmigaBitmapFont(font, file.path(tempdir(), "9"))

## The following examples require the 'adfExplorer' package:
if (requireNamespace("adfExplorer")) {
  library("adfExplorer")
  virtual_disk_file <- tempfile(fileext = ".adf") |>
    create_adf_device(write_protected = FALSE) |>
    prepare_adf_device("font_disk") |>
    make_adf_dir("FONTS")
  
  dest <- virtual_path(virtual_disk_file, "DF0:FONTS")
  write.AmigaBitmapFontSet(font_example, dest)
  close(virtual_disk_file)
}