Write an Amiga Workbench icon (info) file

Graphical representation of files and directories (icons) are stored as separate files (with the .info extension) on the Amiga. This function writes AmigaIcon() class objects to such files.

Usage

write.AmigaIcon(x, file)

Arguments

x

An AmigaIcon() class object.

file

A character string representing the file name to which the icon data should be written.

Value

Returns NULL or an integer status passed on by the close() function, that is used to close the file connection. It is returned invisibly.

Details

The AmigaIcon() S3 object provides a comprehensive format for Amiga icons, which are used as a graphical representation of files and directories on the Amiga. The AmigaIcon() is a named list containing all information of an icon. Use this function to write this object to a file which can be used on the Commodore Amiga or emulator.

See also

Author

Pepijn de Vries

Examples

## create a simple AmigaIcon:
icon <- simpleAmigaIcon()

## write the icon to the temp dir:
write.AmigaIcon(icon, file.path(tempdir(), "icon.info"))