Write an Amiga system-configuration file

Write a SysConfig class object to an Amiga binary system-configuration file.

Usage

write.SysConfig(x, file)

Arguments

x

An S3 SysConfig class object.

file

A file name to which the binary file should be written.

Value

Returns NULL or an integer status passed on by the close() function, that is used to close the file connection. It is returned invisibly.

Details

Amiga OS 1.x stored system preferences in a binary system-configuration file. This function writes a SysConfig class object as such a binary file. This file can be used on an Amiga or in an emulator.

See also

Author

Pepijn de Vries

Examples

## First generate a simple SysConfig object to write to a file:
sc <- simpleSysConfig()

## And write to the tempdir:
write.SysConfig(sc, file.path(tempdir(), "system-configuration"))