Write Interchange File Format (IFF)

Write an IFFChunk() object conform the Interchange File Format (IFF).

Usage

write.iff(x, file)

Arguments

x

An IFFChunk() object that needs to be written to a file.

file

A filename for the IFF file to which the IFFChunk() needs to be saved, or a connection to which the data should be written.

Value

Returns either NULL or an integer status invisibly as passed by the close() statement used to close the file connection.

Details

Writes an IFFChunk() object (including all nested chunks) to the specified file. Only the structure of the object needs to be valid, however, a correctly structured file does not necessarily result in an interpretable file (see examples).

References

https://en.wikipedia.org/wiki/Interchange_File_Format

See also

Author

Pepijn de Vries

Examples

## read an IFF file as an IFFChunk object:
example.iff <- read.iff(system.file("ilbm8lores.iff", package = "AmigaFFH"))

## This will write the IFF file (in this case a bitmap image)
## to the temp directory:
write.iff(example.iff, file.path(tempdir(), "image.iff"))