Coerce raw data into a SysConfig class object

SysConfig objects are comprehensive representations of binary Amiga system-configuration files. Use this function to convert raw data from such a file to a SysConfig object.

Usage

rawToSysConfig(x)

Arguments

x

A vector of raw data that needs to be converted into an S3 SysConfig class object. It should have a length of at least 232. Although system-configurations can be extended, such extended files are not supported here.

Value

Returns a SysConfig class object based on x.

Details

The Amiga used the system-configuration file to store certain system preferences in a binary file. With this function such raw data can be converted into a more comprehensive SysConfig object. Use as.raw() to achieve the inverse.

See also

Author

Pepijn de Vries

Examples

if (FALSE) {
## get the system-configuration from the adfExplorer example disk:
sc <- adfExplorer::get.adf.file(adfExplorer::adf.example, "devs/system-configuration")

## This will get you the raw data from the file:
typeof(sc)

## Convert the raw data to a more comprehensive named list (and S3 SysConfig class):
sc <- rawToSysConfig(sc)
}