Create simple AmigaIcon objects

Graphical representation of files and directories (icons) are stored as separate files (with the .info extension) on the Amiga. This function writes AmigaIcon() class objects to such files.

Usage

simpleAmigaIcon(
  version = c("OS1.x", "OS2.x"),
  type = c("WBDISK", "WBDRAWER", "WBTOOL", "WBPROJECT", "WBGARBAGE", "WBDEVICE",
    "WBKICK", "WBAPPICON"),
  two.images = TRUE,
  back.fill = FALSE,
  ...
)

Arguments

version

A character string indicating the Amiga OS version with which the icon should be compatible. "OS2.x" indicates \>=OS2.0 and "OS1.x" indicates <OS2.0.

type

A character string indicating the type of object (file, disk, directory, etc.) the icon should represent. See the `Usage' section for all posible options.

two.images

A single logical value, indicating whether the selected icon is depicted as a second image (in which case the icon contains two images). The default value is TRUE.

back.fill

A single logical value, indicating whether the selected image of the icon should use the back fill' mode (default). If set to FALSE complement' mode is used. Note that back fill is not compatible when the icon holds two images. In the complement' mode, the image colours are inverted when selected. In the back fill' exterior first colour is not inverted.

...

Reserved for additional arguments. Currently ignored.

Value

Returns a simple S3 object of class AmigaIcon().

Details

This function creates basic AmigaIcon() objects which can be modified afterwards. It uses simple generic images to represent different types of files or directories.

See also

Author

Pepijn de Vries

Examples

if (FALSE) {
## Create an AmigaIcon object using the default arguments:
icon <- simpleAmigaIcon()
}